pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "product-marketing-agency"
version = "1.0.0"
description = "A comprehensive multi-agent marketing system for generating various types of product marketing images and content using the swarms framework"
license = "MIT"
authors = ["The Swarm Corporation <kye@swarms.world>"]
homepage = "https://github.com/The-Swarm-Corporation/Product-Marketing-Agency"
documentation = "https://github.com/The-Swarm-Corporation/Product-Marketing-Agency/blob/main/docs/README.md"
readme = "README.md"
repository = "https://github.com/The-Swarm-Corporation/Product-Marketing-Agency"
keywords = ["marketing", "product photography", "ai agents", "swarms", "image generation", "marketing automation", "multi-agent systems", "content creation", "e-commerce"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: End Users/Desktop",
    "Topic :: Multimedia :: Graphics",
    "Topic :: Office/Business",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12"
]

[tool.poetry.dependencies]
python = "^3.10"
swarms = "*"
rich = "*"
python-dotenv = "*"
loguru = "*"

[tool.poetry.dev-dependencies]
pytest = "^7.0.0"
pytest-cov = "^4.0.0"
black = "^23.1.0"
ruff = "^0.1.6"
mypy = "^1.0.0"

[tool.poetry.group.lint.dependencies]
ruff = "^0.1.6"
types-toml = "^0.10.8.1"
types-redis = "^4.3.21.6"
types-pytz = "^2023.3.0.0"
black = "^23.1.0"
types-chardet = "^5.0.4.6"
mypy-protobuf = "^3.0.0"

[tool.autopep8]
max_line_length = 80
ignore = "E501,W6"
in-place = true
recursive = true
aggressive = 3

[tool.ruff]
line-length = 70

[tool.black]
line-length = 70
target-version = ['py310']
preview = true